\name{testIntNA-package}
\alias{testIntNA-package}
\alias{testIntNA}
\docType{package}
\title{
\packageTitle{testIntNA}
}
\description{
\packageDescription{testIntNA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{testIntNA}
\packageIndices{testIntNA}
bla bla
}
\author{
\packageAuthor{testIntNA}

Maintainer: \packageMaintainer{testIntNA}
}
\references{
Nothing here
}
\keyword{ package }
\seealso{
TMB
}
\examples{
2+2
}
